moved data crate to queue, refactored message push and retrieve
